Edit: ip 차단 함수를 loaders 폴더로 이동
